Adjust pointers into flyweight during tsdp_qname_dup
Otherwise, the duplicate has copies of pointers that point to the
original flyweight.  If you then tsdp_qname_free the original, you are
left with dangling pointers and a subsequent tsdp_qname_string will
segfault.

The tests now dup the qname and then free the original to tickle this
particular not-so-edge-case.